+++ crates/renderer/src/renderer.rs
@@ -18,10 +18,6 @@ use freya_elements::events::{
     Code,
     Key,
 };
-use glutin::prelude::{
-    GlSurface,
-    PossiblyCurrentGlContext,
-};
 use torin::geometry::CursorPoint;
 use winit::{
     application::ApplicationHandler,
@@ -45,9 +41,7 @@ use winit::{
 
 use crate::{
     devtools::Devtools,
-    size::WinitSize,
     window_state::{
-        create_surface,
         CreatedState,
         NotCreatedState,
         WindowState,
@@ -243,18 +237,13 @@ impl<'a, State: Clone> ApplicationHandler<EventMessage> for DesktopRenderer<'a, 
     ) {
         let scale_factor = self.scale_factor();
         let CreatedState {
-            gr_context,
             surface,
             dirty_surface,
-            gl_surface,
-            gl_context,
             window,
             window_config,
             app,
-            fb_info,
-            num_samples,
-            stencil_size,
             is_window_focused,
+            graphics_driver,
             ..
         } = self.state.created_state();
         app.accessibility
@@ -287,7 +276,7 @@ impl<'a, State: Clone> ApplicationHandler<EventMessage> for DesktopRenderer<'a, 
 
                     app.measure_layout_on_next_render = false;
                 }
-                gl_context.make_current(gl_surface).unwrap();
+                graphics_driver.make_current();
                 app.render(
                     &self.hovered_node,
                     window_config.background,
@@ -298,8 +287,7 @@ impl<'a, State: Clone> ApplicationHandler<EventMessage> for DesktopRenderer<'a, 
 
                 app.event_loop_tick();
                 window.pre_present_notify();
-                gr_context.flush_and_submit();
-                gl_surface.swap_buffers(gl_context).unwrap();
+                graphics_driver.flush_and_submit();
             }
             WindowEvent::MouseInput { state, button, .. } => {
                 app.set_navigation_mode(NavigationMode::NotKeyboard);
@@ -422,12 +410,10 @@ impl<'a, State: Clone> ApplicationHandler<EventMessage> for DesktopRenderer<'a, 
                 });
             }
             WindowEvent::Resized(size) => {
-                *surface =
-                    create_surface(window, *fb_info, gr_context, *num_samples, *stencil_size);
+                let (new_surface, new_dirty_surface) = graphics_driver.resize(size);
 
-                *dirty_surface = surface.new_surface_with_dimensions(size.to_skia()).unwrap();
-
-                gl_surface.resize(gl_context, size.as_gl_width(), size.as_gl_height());
+                *surface = new_surface;
+                *dirty_surface = new_dirty_surface;
 
                 window.request_redraw();
 
@@ -459,21 +445,5 @@ impl<'a, State: Clone> ApplicationHandler<EventMessage> for DesktopRenderer<'a, 
 
     fn exiting(&mut self, _event_loop: &winit::event_loop::ActiveEventLoop) {
         self.run_on_exit();
-    }
-}
-
-impl<T: Clone> Drop for DesktopRenderer<'_, T> {
-    fn drop(&mut self) {
-        if let WindowState::Created(CreatedState {
-            gl_context,
-            gl_surface,
-            gr_context,
-            ..
-        }) = &mut self.state
-        {
-            if !gl_context.is_current() && gl_context.make_current(gl_surface).is_err() {
-                gr_context.abandon();
-            }
-        }
     }
 }
